check-ignore: fix handling with negated patterns
check-ignore was meant to check ignore rules the same way git status and
other commands would, and to report whether a path is excluded.  It
failed to do this (and generated a few bug reports), however, because it
did not account for negated patterns.

Commands other than check-ignore verify exclusion rules via calling

   ... -> treat_one_path() -> is_excluded() -> last_matching_pattern()

while check-ignore has a call path of the form:

   ... -> check_ignore()                    -> last_matching_pattern()

The fact that the latter does not include the call to is_excluded()
means that it is susceptible to to messing up negated patterns (since
that is the only significant thing is_excluded() adds over
last_matching_pattern()).  Unfortunately, we can't make it just call
is_excluded(), because is_excluded doesn't return the pattern in
question and part of check-ignore's functionality is not just checking
whether one of the patterns matches but returning which one does.

Further, check_ignore() is supposed to handle a --verbose mode, which
was ill-defined for the case of negated patterns: check-ignore was
documented to print just the excluded paths, whereas the --verbose mode
was there to document which patterns were matched by paths.  A path
which matches a negated exclude pattern is NOT excluded and thus
shouldn't be printed by the former logic, while it certainly does match
one of the explicit patterns and thus should be printed by the latter
logic.  Adjust the definition of --verbose to state that it is about
matching patterns INSTEAD of about showing which paths are excluded in
order to resolve this discrepancy.

Finally, also adjust a few tests in t0008 that were caught up by this
discrepancy in how negated paths were handled.

Signed-off-by: Elijah Newren <newren@gmail.com>