describe: output tag's ref instead of embedded name
If a tag describes a commit, we currently output not the tag's ref but
its embedded name. This means that when the tag is locally stored under
a different name, the output given cannot be used to access the tag in
any way. A warning is also emitted in this case, but the message is not
very enlightening:

$ git tag -am "testing tag body" testing-tag
$ mv .git/refs/tags/testing-tag .git/refs/tags/testing-tag-with-new-name
$ git describe --tags --abbrev=0
warning: tag 'testing-tag' is really 'testing-tag-with-new-name' here
testing-tag

Let's make git-describe output the tag's local name instead and
rephrase the warning to reflect the situation a little better.

Since the embedded name will no longer be needed for correct output, we
can convert the die() call in append_name() when a tag doesn't have the
embedded name to a warning(). Also, this function will now have two
disconnected responsibilities: verifying if the tag's embedded name
matches the ref and actually appending the ref to a given buffer (which
does not depend on the parsed tag object itself). Thus, to increase
intelligibility, let's make the function specialize in the former and do
the latter outside it.

Suggested-by: Jeff King <peff@peff.net>
Suggested-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>
Acked-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>