rebase: do not switch the default to 'merge' just yet
Revert 2ac0d627 (rebase: change the default backend from "am" to
"merge", 2020-02-15) to postpone the switch of default backend of
"git rebase" to the merge backend, as there seem to be a few
remaining bugs (we saw two reported on the day after 2.26-rc1 was
released---we do not know how many remaining bugs there are) that
regress the end user experience.

 * When a rebase stops with a merge conflict, "rebase --continue"
   after resolving the conflict opens an editor with the merge
   backend.

 * When a rebase sees a change that is already applied, the end user
   gets thrown into "cherry-picking" mode, causing "git status" to
   say "nothing to commit, working tree clean". At that point, "git
   rebase --continue" does not let the user get out of this state.

Let's keep the default for the upcoming release, without removing
the configuration variable so that those adventurous can opt into
using the 'merge' backend to help polish it.

Acked-by: Elijah Newren <newren@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>