sequencer: clear CHERRY_PICK_HEAD upon dropping a become-empty commit
In commit e98c4269c8 ("rebase (interactive-backend): fix handling of
commits that become empty", 2020-02-15), the merge backend was changed
to drop commits that did not start empty but became so after being
applied (because their changes were a subset of what was already
upstream).  This new code path did not need to go through the process of
creating a commit, since we were dropping the commit instead.
Unfortunately, this also means we bypassed the clearing of the
CHERRY_PICK_HEAD file, which if there were no further commits to
cherry-pick would mean that the rebase would end but assume there was
still an operation in progress.  Ensure that we clear CHERRY_PICK_HEAD
when we decide to drop the commit.

Signed-off-by: Elijah Newren <newren@gmail.com>