doc: log takes multiple revision ranges
`git log` accepts multiple revision ranges, but makes no mention of that
in its documentation. Currently readers have to go to the `git
rev-list` documentation to discover this useful feature.

Mention it and adjust the synopsys in the doc and in the short
help for `git log`.

Also add links to the documentation for `git rev-list` and `git diff`
instead of simply mentioning them, to make it easier for readers to reach
these documentation pages. Let's link to `git diff` as this is the
porcelain command, and the rest of the family (`diff-index`, `diff-tree` and
`diff-files`) are mentioned in the "Raw output format" section of the
`git diff` documentation.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>